Edit: Optimize file structures
Reorder the file structure for more clean use. Create a components folder, that only has resolutions folder. In there are two files: 'resolution.js' and 'resolution.html', in the js only import the client side helpers and events for template resolution and in the html import the template resolution. Renamed the 'resoñutions.html' to layout.html for more accurate name and understandable.